Fixing my dumb git mistakes (#17)
* Fixing my dumb git mistakes

* remove unwanted file
