--- matcha.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "backend", "backend\backend.csproj", "{769E809C-7A97-4658-8311-FEAC2C64AD85}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{769E809C-7A97-4658-8311-FEAC2C64AD85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{769E809C-7A97-4658-8311-FEAC2C64AD85}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{769E809C-7A97-4658-8311-FEAC2C64AD85}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{769E809C-7A97-4658-8311-FEAC2C64AD85}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {158F39A6-FA80-4F0A-BE3B-F2BD59642767}
-	EndGlobalSection
-EndGlobal
